backend/requirements.txt
flask
flask-cors
googlemaps
python-dotenv
langchain-core
langchain
langchain-community
langchain-google-genai
langgraph
pydantic
requests
pymongo
bson
sympy
aiohttp
httpx
google-cloud-translate
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
